BOOK :white_check_mark: add snapshot tests for CarouselSlide
in the process remove a lot of the plain assertions on the component
question is if you needed to refactor this functionality out would
the original test help you more or is the snapshot ok? enough?
how easy is it to follow? book assumes pull request process and review
but we are talking trunk based development on master.