Attempt to deal with API errors and parse json
The issue is that the apilayer.com service seems to be super flakey and
often cloudflare seems to kick in returning a 524 with no body which
fails the error parsing. Add some way to deal with that? also would be
good to timeout as this takes 1minute to timeout by default?

although there seems to be some content-length? 15? and the report-to
header seems to have some cloudflare url with a token?

result:
    Response {
        url: Url {
            scheme: "https",
            host: Some(
                Domain("api.apilayer.com"),
            ),
            path: "/bad_words",
            query: Some("censor_character=*"),
        },
        status: 524,
        headers: {
            "date": "Fri, 14 Jun 2024 12:32:41 GMT",
            "content-type": "text/plain; charset=UTF-8",
            "content-length": "15",
            "connection": "keep-alive",
            "report-to": "{\"endpoints\":
            [{\"url\":\"https:\/\/a.nel.cloudflare.com\/report\/v4?s=5eJ...zTV\"}],
            \"group\":\"cf-nel\",\"max_age\":604800}",
        },
    }
    thread 'tokio-runtime-worker' panicked at src/routes/question.rs:138:50:
    called `Result::unwrap()` on an `Err` value: reqwest::Error { kind:
        Decode, source: Error("expected value", line: 1, column: 1) }
    note: run with `RUST_BACKTRACE=1` environment variable to display a
        backtrace