[IONIC] RecipeService and display a recipe
with:
    ionic generate service recipes/recipes

also note that to deal with the "string | null" or no constructor added
a tsconfig property
    "strictPropertyInitialization": false,

also had a bunch of issues with typescript NOT being able to identify
that a value was truthy so used the ! to prompt it. Also used
    `as Recipe`
to deal with typescript thinking it may have undefined values.