[IONIC] add more navigation from list to deets
Added a service and a model (manually) with:
    ionic generate service places/places --skip-tests

The service for the time being stores a bunch of hard coded data. The
navigation on Discover is implemented to visit a specific place.

There was also discussion that a RouterLink can have a direction
forward and backward to make sure the transition styling is correct but
it can usually do the right thing by default.
eg:
    <ion-button
      routerLink="/location"
      direction="forward"
      // OR
      direction="backward"
    >